requirements.txt
boto3==1.4.4.
celery==3.1.23
cffi==1.9.1
cryptography==1.7.1
Django==1.9.2
django-avatar==3.1.0
django-celery==3.1.17
django-storages==1.5.2
ecdsa==0.13
Fabric==1.13.1
gevent==1.2.1
gunicorn==19.6.0
idna==2.2
ipaddress==1.0.18
kombu==3.0.35
paramiko==2.1.1
psycopg2==2.7.1
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
PyJWT==1.4.2
pyrebase==3.0.27
redis==2.10.5
requests==2.11.1
requests-toolbelt==0.7.0
six==1.10.0
